Added support for get/sendPartialObject
If the caps flags are enabled for both get and send of partial object
this method is used instead of file push and pull. This makes many file
operations much faster and is widely supported throughout mobile
devices.

We still use the tmp file, not for storing data but to keep a unique
file handler for all the fuse file operations.

No caching happens at the moment, we could cache in the tmp file and
keep a table of offsets and sizes of the file that we already cached,
however, using the kernel file system cache is much better.